apps/api/requirements.txt
coverage==7.4.1
django==5.1.3
djangorestframework==3.15.2
django-filter==23.5
djangorestframework-simplejwt[crypto]==5.3.1
djangorestframework-gis==1.0
django-auditlog==3.0.0
drf-access-policy==1.5.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.1.1
flake8==7.0.0
Faker==22.5.0
phonenumbers==8.13.28
psycopg==3.1.17
psycopg-c==3.1.17
psycopg-pool==3.2.1
python-decouple==3.8
gunicorn==21.2.0
requests==2.31.0
resend==0.7.2
django-cors-headers==4.3.1
arrow==1.3.0
django-tree-queries==0.16.1
pika==1.3.2
drf-standardized-errors[openapi]==0.13.0
rest-framework-generic-relations==2.2.0
pillow==10.4.0
pyseaweed==2.0.0
django-sequences==3.0
carmen-cloud-client==1.2.0
humanize==4.10.0
freezegun==1.5.1
django-two-factor-auth==1.17.0
django-axes==6.5.1
sentry-sdk==2.13.0
transitions==0.9.2
django-soft-delete==1.0.16